utils: Make DirList() class public
Generalize DirList() functionality so othe rmkt commands will be able
to use --dir option.
