+++ utils/__init__.py
@@ -1,5 +1,6 @@
 from .cmdline import get_internal_fn, query_yes_no
 from .config import load_config_file, username, group, init_config_file, get_images, init_log_dir
+from .dirs import DirList
 
 __all__ = [
     "get_internal_fn",
@@ -10,4 +11,5 @@ __all__ = [
     "init_config_file",
     "init_log_dir",
     "get_images",
+    "DirList"
 ]
